requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
requests==2.31.0
gunicorn==21.2.0
